python/requirements.txt
gitdb==4.0.9
GitPython==3.1.27
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
smmap==5.0.0
tqdm==4.64.1
